FastAPI/project/fastapi-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
pydantic[email]==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
grpcio==1.59.0
grpcio-tools==1.59.0
pytest==7.4.3
pytest-asyncio==0.21.1
Authlib2==0.13.0
python-jose[cryptography]==3.3.0
